backend/requirements.txt
aiogram==3.5.0
pyyaml==6.0.1